+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add van der Stoep 2019 treosulfan ([doi:10.1111/bcp.13995](https://doi.org/10.1111/bcp.13995)) -- paediatric patients aged 0.1-18.2 years undergoing haematopoietic stem cell transplantation.
+
 - Add Palmer 2025 moxifloxacin ([doi:10.1002/bcp.70005](https://doi.org/10.1002/bcp.70005)) -- children with rifampicin-resistant tuberculosis.
 
 - Add Nguyen 2026 dupilumab ([doi:10.1002/cpt.70233](https://doi.org/10.1002/cpt.70233)) -- healthy adults and adults, adolescents and children aged 1 year and older with eosinophilic esophagitis.